requirements.txt
# LakeFusion Unified App - Merged Requirements
# Generated by scripts/build_lakefusion_app.sh

# LakeFusion utility (built from source)
./wheels/lakefusion_utility-1.0.0-py3-none-any.whl

aioredis==1.3.1
alembic==1.13.2
APScheduler==3.10.4
beautifulsoup4==4.13.4
cryptography==44.0.1
databricks-sdk==0.56.0
databricks-sql-connector[pyarrow]==4.2.3
databricks-vectorsearch==0.60
fastapi==0.124.0
fastapi-cache==0.1.0
fastapi-cache2==0.2.2
fastmcp==3.2.0
google-auth==2.36.0
hiredis==3.0.0
httpx>=0.27.0
lz4==4.4.5
marshmallow==4.1.2
mlflow-skinny==3.4.0
numpy==2.3.5
oauthlib==3.2.2
openpyxl==3.1.5
pandas==2.3.3
pendulum==3.0.0
psycopg2-binary==2.9.11
pycryptodome==3.23.0
pydantic==2.12.5
PyMySQL==1.1.1
pyodbc
pytest>=8.0.0
pytest-asyncio>=0.24.0
python-multipart==0.0.22
requests==2.32.4
sendgrid==6.11.0
SQLAlchemy==2.0.34
starlette==0.49.1
thrift==0.20.0
time-machine==2.16.0
uvicorn==0.35
